separated server and client components. added suspense
